requirements.txt
Flask==2.3.3
mysql-connector-python==8.1.0
Jinja2==3.1.2
Werkzeug==2.3.7
pycryptodome==3.18.0